[update] b-15 関数置き場所移動
