+++ src/index.js
@@ -3,5 +3,6 @@ import * as date from './date';
 import * as image from './image';
 import * as url from './url';
 import * as http from './http';
+import * as collection from './collection';
 
-export {array, date, image, url, http};
+export {array, date, image, url, http, collection};
